move type definition
