+++ homeassistant/components/tedee/coordinator.py
@@ -1,5 +1,7 @@
 """Coordinator for Tedee locks."""
 
+from __future__ import annotations
+
 from collections.abc import Awaitable, Callable
 from datetime import timedelta
 import logging
@@ -31,6 +33,8 @@ GET_LOCKS_INTERVAL_SECONDS = 3600
 
 _LOGGER = logging.getLogger(__name__)
 
+type TedeeConfigEntry = ConfigEntry[TedeeApiCoordinator]
+
 
 class TedeeApiCoordinator(DataUpdateCoordinator[dict[int, TedeeLock]]):
     """Class to handle fetching data from the tedee API centrally."""
@@ -38,7 +42,7 @@ class TedeeApiCoordinator(DataUpdateCoordinator[dict[int, TedeeLock]]):
     config_entry: ConfigEntry
     bridge: TedeeBridge
 
-    def __init__(self, hass: HomeAssistant, entry: ConfigEntry) -> None:
+    def __init__(self, hass: HomeAssistant, entry: TedeeConfigEntry) -> None:
         """Initialize coordinator."""
         super().__init__(
             hass,
