+++ homeassistant/components/tedee/sensor.py
@@ -15,7 +15,7 @@ from homeassistant.const import PERCENTAGE, EntityCategory, UnitOfTime
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import TedeeConfigEntry
+from .coordinator import TedeeConfigEntry
 from .entity import TedeeDescriptionEntity
